backend/requirements.txt
# Django Core
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Redis & Celery
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
cryptography>=41.0.0,<42.0.0

# API Documentation
drf-spectacular==0.26.5

# File handling & Storage
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0

# AI/OCR Dependencies
pytesseract==0.3.10
easyocr==1.7.0
opencv-python==4.8.1.78
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
numpy==1.24.4

# Development
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2 